Change from BuildConfig to OAuthProviderSettingsFactory
