+++ app/k9mail/build.gradle.kts
@@ -87,28 +87,6 @@ android {
                 "proguard-rules.pro",
             )
 
-            buildConfigField(
-                "String",
-                "OAUTH_GMAIL_CLIENT_ID",
-                "\"262622259280-hhmh92rhklkg2k1tjil69epo0o9a12jm.apps.googleusercontent.com\"",
-            )
-            buildConfigField(
-                "String",
-                "OAUTH_YAHOO_CLIENT_ID",
-                "\"dj0yJmk9aHNUb3d2MW5TQnpRJmQ9WVdrOWVYbHpaRWM0YkdnbWNHbzlNQT09JnM9Y29uc3VtZXJzZWNyZXQmc3Y9MCZ4PWIz\"",
-            )
-            buildConfigField(
-                "String",
-                "OAUTH_AOL_CLIENT_ID",
-                "\"dj0yJmk9dUNqYXZhYWxOYkdRJmQ9WVdrOU1YQnZVRFZoY1ZrbWNHbzlNQT09JnM9Y29uc3VtZXJzZWNyZXQmc3Y9MCZ4PWIw\"",
-            )
-            buildConfigField("String", "OAUTH_MICROSOFT_CLIENT_ID", "\"e647013a-ada4-4114-b419-e43d250f99c5\"")
-            buildConfigField(
-                "String",
-                "OAUTH_MICROSOFT_REDIRECT_URI_ID",
-                "\"Dx8yUsuhyU3dYYba1aA16Wxu5eM%3D\"",
-            )
-
             manifestPlaceholders["appAuthRedirectScheme"] = "com.fsck.k9"
         }
 
@@ -118,28 +96,6 @@ android {
             enableAndroidTestCoverage = testCoverageEnabled
 
             isMinifyEnabled = false
-
-            buildConfigField(
-                "String",
-                "OAUTH_GMAIL_CLIENT_ID",
-                "\"262622259280-5qb3vtj68d5dtudmaif4g9vd3cpar8r3.apps.googleusercontent.com\"",
-            )
-            buildConfigField(
-                "String",
-                "OAUTH_YAHOO_CLIENT_ID",
-                "\"dj0yJmk9ejRCRU1ybmZjQlVBJmQ9WVdrOVVrZEViak4xYmxZbWNHbzlNQT09JnM9Y29uc3VtZXJzZWNyZXQmc3Y9MCZ4PTZj\"",
-            )
-            buildConfigField(
-                "String",
-                "OAUTH_AOL_CLIENT_ID",
-                "\"dj0yJmk9cHYydkJkTUxHcXlYJmQ9WVdrOWVHZHhVVXN4VVV3bWNHbzlNQT09JnM9Y29uc3VtZXJzZWNyZXQmc3Y9MCZ4PTdm\"",
-            )
-            buildConfigField("String", "OAUTH_MICROSOFT_CLIENT_ID", "\"e647013a-ada4-4114-b419-e43d250f99c5\"")
-            buildConfigField(
-                "String",
-                "OAUTH_MICROSOFT_REDIRECT_URI_ID",
-                "\"VZF2DYuLYAu4TurFd6usQB2JPts%3D\"",
-            )
 
             manifestPlaceholders["appAuthRedirectScheme"] = "com.fsck.k9.debug"
         }
